+++ packages/server/src/Hocuspocus.ts
@@ -1,14 +1,14 @@
 import { IncomingMessage } from 'http'
-import { ListenOptions } from 'net'
+import { fileURLToPath } from 'node:url'
+import fs from 'node:fs'
+import { dirname, join } from 'node:path'
 import {
   ResetConnection, awarenessStatesToArray,
 } from '@hocuspocus/common'
-import kleur from 'kleur'
 import { v4 as uuid } from 'uuid'
-import WebSocket, { AddressInfo } from 'ws'
+import WebSocket from 'ws'
 import { Doc, applyUpdate, encodeStateAsUpdate } from 'yjs'
-import meta from '../package.json' assert { type: 'json' }
-import { Server as HocuspocusServer } from './Server.js'
+import { Server } from './Server.js'
 import { ClientConnection } from './ClientConnection.js'
 // TODO: would be nice to only have a dependency on ClientConnection, and not on Connection
 import Connection from './Connection.js'
@@ -24,16 +24,15 @@ import {
   beforeBroadcastStatelessPayload,
   onChangePayload,
   onDisconnectPayload,
-  onListenPayload,
   onStoreDocumentPayload,
 } from './types.js'
 import { getParameters } from './util/getParameters.js'
 import { useDebounce } from './util/debounce.js'
 
+const meta = JSON.parse(fs.readFileSync(join(dirname(fileURLToPath(import.meta.url)), '../package.json'), 'utf-8'))
+
 export const defaultConfiguration = {
   name: null,
-  port: 80,
-  address: '0.0.0.0',
   timeout: 30000,
   debounce: 2000,
   maxDebounce: 10000,
@@ -43,12 +42,8 @@ export const defaultConfiguration = {
     gcFilter: () => true,
   },
   unloadImmediately: true,
-  stopOnSignals: true,
 }
 
-/**
- * Hocuspocus Server
- */
 export class Hocuspocus {
   configuration: Configuration = {
     ...defaultConfiguration,
@@ -76,7 +71,7 @@ export class Hocuspocus {
 
   documents: Map<string, Document> = new Map()
 
-  server?: HocuspocusServer
+  server?: Server
 
   debugger = new Debugger()
 
@@ -89,7 +84,7 @@ export class Hocuspocus {
   }
 
   /**
-   * Configure the server
+   * Configure Hocuspocus
    */
   configure(configuration: Partial<Configuration>): Hocuspocus {
     this.configuration = {
@@ -150,119 +145,6 @@ export class Hocuspocus {
   }
 
   /**
-   * Start the server
-   */
-  async listen(
-    portOrCallback: number | ((data: onListenPayload) => Promise<any>) | null = null,
-    callback: any = null,
-  ): Promise<Hocuspocus> {
-    if (typeof portOrCallback === 'number') {
-      this.configuration.port = portOrCallback
-    }
-
-    if (typeof portOrCallback === 'function') {
-      this.configuration.extensions.push({
-        onListen: portOrCallback,
-      })
-    }
-
-    if (typeof callback === 'function') {
-      this.configuration.extensions.push({
-        onListen: callback,
-      })
-    }
-
-    this.server = new HocuspocusServer(this)
-
-    if (this.configuration.stopOnSignals) {
-      const signalHandler = async () => {
-        await this.destroy()
-        process.exit(0)
-      }
-
-      process.on('SIGINT', signalHandler)
-      process.on('SIGQUIT', signalHandler)
-      process.on('SIGTERM', signalHandler)
-    }
-
-    return new Promise((resolve: Function, reject: Function) => {
-      this.server?.httpServer.listen({
-        port: this.configuration.port,
-        host: this.configuration.address,
-      } as ListenOptions, async () => {
-        if (!this.configuration.quiet && process.env.NODE_ENV !== 'testing') {
-          this.showStartScreen()
-        }
-
-        const onListenPayload = {
-          instance: this,
-          configuration: this.configuration,
-          port: this.address.port,
-        }
-
-        try {
-          await this.hooks('onListen', onListenPayload)
-          resolve(this)
-        } catch (e) {
-          reject(e)
-        }
-      })
-    })
-  }
-
-  get address(): AddressInfo {
-    return (this.server?.httpServer?.address() || {
-      port: this.configuration.port,
-      address: this.configuration.address,
-      family: 'IPv4',
-    }) as AddressInfo
-  }
-
-  get URL(): string {
-    return `${this.configuration.address}:${this.address.port}`
-  }
-
-  get webSocketURL(): string {
-    return `ws://${this.URL}`
-  }
-
-  get httpURL(): string {
-    return `http://${this.URL}`
-  }
-
-  private showStartScreen() {
-    const name = this.configuration.name ? ` (${this.configuration.name})` : ''
-
-    console.log()
-    console.log(`  ${kleur.cyan(`Hocuspocus v${meta.version}${name}`)}${kleur.green(' running at:')}`)
-    console.log()
-    console.log(`  > HTTP: ${kleur.cyan(`${this.httpURL}`)}`)
-    console.log(`  > WebSocket: ${this.webSocketURL}`)
-
-    const extensions = this.configuration?.extensions.map(extension => {
-      return extension.extensionName ?? extension.constructor?.name
-    })
-      .filter(name => name)
-      .filter(name => name !== 'Object')
-
-    if (!extensions.length) {
-      return
-    }
-
-    console.log()
-    console.log('  Extensions:')
-
-    extensions
-      .forEach(name => {
-        console.log(`  - ${name}`)
-      })
-
-    console.log()
-    console.log(`  ${kleur.green('Ready.')}`)
-    console.log()
-  }
-
-  /**
    * Get the total number of active documents
    */
   getDocumentsCount(): number {
@@ -273,10 +155,17 @@ export class Hocuspocus {
    * Get the total number of active connections
    */
   getConnectionsCount(): number {
-    return Array.from(this.documents.values()).reduce((acc, document) => {
-      acc += document.getConnectionsCount()
-      return acc
+    const uniqueSocketIds = new Set<string>()
+    const totalDirectConnections = Array.from(this.documents.values()).reduce((acc, document) => {
+      // Accumulate unique socket IDs
+      document.getConnections().forEach(({ socketId }) => {
+        uniqueSocketIds.add(socketId)
+      })
+      // Accumulate direct connections
+      return acc + document.directConnectionsCount
     }, 0)
+    // Return the sum of unique socket IDs and direct connections
+    return uniqueSocketIds.size + totalDirectConnections
   }
 
   /**
@@ -296,38 +185,6 @@ export class Hocuspocus {
         connection.close(ResetConnection)
       })
     })
-  }
-
-  /**
-   * Destroy the server
-   */
-  async destroy(): Promise<any> {
-    await new Promise(async resolve => {
-
-      this.server?.httpServer?.close()
-
-      try {
-
-        this.configuration.extensions.push({
-          async afterUnloadDocument({ instance }) {
-            if (instance.getDocumentsCount() === 0) resolve('')
-          },
-        })
-
-        this.server?.webSocketServer?.close()
-        if (this.getDocumentsCount() === 0) resolve('')
-
-        this.closeConnections()
-
-      } catch (error) {
-        console.error(error)
-      }
-
-      this.debugger.flush()
-
-    })
-
-    await this.hooks('onDestroy', { instance: this })
   }
 
   /**
@@ -392,10 +249,7 @@ export class Hocuspocus {
       transactionOrigin: connection,
     }
 
-    this.hooks('onChange', hookPayload).catch(error => {
-      // TODO: what's the intention of this catch -> throw?
-      throw error
-    })
+    this.hooks('onChange', hookPayload)
 
     // If the update was received through other ways than the
     // WebSocket connection, we don’t need to feel responsible for
@@ -522,14 +376,14 @@ export class Hocuspocus {
       () => {
         return this.hooks('onStoreDocument', hookPayload)
           .then(() => {
-            this.hooks('afterStoreDocument', hookPayload).then(() => {
+            this.hooks('afterStoreDocument', hookPayload).then(async () => {
               // Remove document from memory.
 
               if (document.getConnectionsCount() > 0) {
                 return
               }
 
-              this.unloadDocument(document)
+              await this.unloadDocument(document)
             })
           })
           .catch(error => {
@@ -581,13 +435,19 @@ export class Hocuspocus {
     return chain
   }
 
-  unloadDocument(document: Document) {
+  async unloadDocument(document: Document): Promise<any> {
     const documentName = document.name
     if (!this.documents.has(documentName)) return
 
+    await this.hooks('beforeUnloadDocument', { instance: this, documentName })
+
+    if (document.getConnectionsCount() > 0) {
+      return
+    }
+
     this.documents.delete(documentName)
     document.destroy()
-    this.hooks('afterUnloadDocument', { instance: this, documentName })
+    await this.hooks('afterUnloadDocument', { instance: this, documentName })
   }
 
   enableDebugging() {
@@ -635,5 +495,3 @@ export class Hocuspocus {
     return new DirectConnection(document, this, context)
   }
 }
-
-export const Server = new Hocuspocus()