+++ packages/server/src/types.ts
@@ -53,6 +53,7 @@ export interface Extension {
   onAwarenessUpdate?(data: onAwarenessUpdatePayload): Promise<any>;
   onRequest?(data: onRequestPayload): Promise<any>;
   onDisconnect?(data: onDisconnectPayload): Promise<any>;
+  beforeUnloadDocument?(data: beforeUnloadDocumentPayload): Promise<any>;
   afterUnloadDocument?(data: afterUnloadDocumentPayload): Promise<any>;
   onDestroy?(data: onDestroyPayload): Promise<any>;
 }
@@ -76,6 +77,7 @@ export type HookName =
   'onAwarenessUpdate' |
   'onRequest' |
   'onDisconnect' |
+  'beforeUnloadDocument' |
   'afterUnloadDocument' |
   'onDestroy'
 
@@ -99,8 +101,10 @@ export type HookPayloadByName = {
   onRequest: onRequestPayload,
   onDisconnect: onDisconnectPayload,
   afterUnloadDocument: afterUnloadDocumentPayload,
+  beforeUnloadDocument: beforeUnloadDocumentPayload,
   onDestroy: onDestroyPayload,
 }
+
 export interface Configuration extends Extension {
   /**
    * A name for the instance, used for logging.
@@ -110,14 +114,6 @@ export interface Configuration extends Extension {
    * A list of hocuspocus extensions.
    */
   extensions: Array<Extension>,
-  /**
-   * The port which the server listens on.
-   */
-  port?: number,
-  /**
-   * The address which the server listens on.
-   */
-  address?: string,
   /**
    * Defines in which interval the server sends a ping, and closes the connection when no pong is sent back.
    */
@@ -143,13 +139,6 @@ export interface Configuration extends Extension {
    * your onStoreDocument is rate-limited.
    */
   unloadImmediately: boolean,
-
-  /**
-   * the server will gracefully stop if SIGINT, SIGQUIT or SIGTERM is received.
-   *
-   * Set this to false if you don't want that.
-   */
-  stopOnSignals: boolean,
 
   /**
    * options to pass to the ydoc document
@@ -368,6 +357,11 @@ export interface onConfigurePayload {
 }
 
 export interface afterUnloadDocumentPayload {
+  instance: Hocuspocus;
+  documentName: string;
+}
+
+export interface beforeUnloadDocumentPayload {
   instance: Hocuspocus;
   documentName: string;
 }