+++ tests/utils/newHocuspocusProvider.ts
@@ -1,6 +1,6 @@
 import {
   HocuspocusProvider,
-  HocuspocusProviderConfiguration, HocuspocusProviderWebsocketConfiguration,
+  HocuspocusProviderConfiguration, HocuspocusProviderWebsocket, HocuspocusProviderWebsocketConfiguration,
 } from '@hocuspocus/provider'
 import { Hocuspocus } from '@hocuspocus/server'
 import { newHocuspocusProviderWebsocket } from './newHocuspocusProviderWebsocket.js'
@@ -9,9 +9,10 @@ export const newHocuspocusProvider = (
   server: Hocuspocus,
   options: Partial<HocuspocusProviderConfiguration> = {},
   websocketOptions: Partial<HocuspocusProviderWebsocketConfiguration> = {},
+  websocketProvider?: HocuspocusProviderWebsocket,
 ): HocuspocusProvider => {
   return new HocuspocusProvider({
-    websocketProvider: newHocuspocusProviderWebsocket(server, websocketOptions),
+    websocketProvider: websocketProvider ?? newHocuspocusProviderWebsocket(server, websocketOptions),
     // Just use a generic document name for all tests.
     name: 'hocuspocus-test',
     // There is no need to share data with other browser tabs in the testing environment.